Cleanups before merge of functions work
- rename confusing FunctionAsExpressionTemplate to VarArgsFunctionDescriptor
- delete an uncalled constructor
- delete an unnecessary class
