+++ hibernate-core/src/main/java/org/hibernate/query/sqm/function/NamedSqmFunctionDescriptor.java
@@ -6,10 +6,15 @@
  */
 package org.hibernate.query.sqm.function;
 
+import java.util.List;
 import java.util.Locale;
 
+import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.query.sqm.produce.function.ArgumentsValidator;
 import org.hibernate.query.sqm.produce.function.FunctionReturnTypeResolver;
+import org.hibernate.sql.ast.SqlAstWalker;
+import org.hibernate.sql.ast.spi.SqlAppender;
+import org.hibernate.sql.ast.tree.SqlAstNode;
 
 /**
  * Provides a standard implementation that supports the majority of the HQL
@@ -20,21 +25,9 @@ import org.hibernate.query.sqm.produce.function.FunctionReturnTypeResolver;
  * @author David Channon
  * @author Steve Ebersole
  */
-public class NamedSqmFunctionDescriptor extends AbstractSqmFunctionDescriptor {
+public class NamedSqmFunctionDescriptor extends AbstractSqmFunctionDescriptor implements FunctionRenderingSupport {
 	private final String functionName;
-	private final FunctionRenderingSupport renderingSupport;
-
-	public NamedSqmFunctionDescriptor(
-			String functionName,
-			boolean useParenthesesWhenNoArgs,
-			ArgumentsValidator argumentsValidator) {
-		this(
-				functionName,
-				useParenthesesWhenNoArgs,
-				argumentsValidator,
-				(impliedTypeAccess, arguments) -> impliedTypeAccess.get()
-		);
-	}
+	private final boolean useParenthesesWhenNoArgs;
 
 	public NamedSqmFunctionDescriptor(
 			String functionName,
@@ -44,7 +37,7 @@ public class NamedSqmFunctionDescriptor extends AbstractSqmFunctionDescriptor {
 		super( argumentsValidator, returnTypeResolver );
 
 		this.functionName = functionName;
-		this.renderingSupport = new StandardFunctionRenderingSupport( useParenthesesWhenNoArgs );
+		this.useParenthesesWhenNoArgs = useParenthesesWhenNoArgs;
 	}
 
 	public String getFunctionName() {
@@ -58,7 +51,35 @@ public class NamedSqmFunctionDescriptor extends AbstractSqmFunctionDescriptor {
 
 	@Override
 	public FunctionRenderingSupport getRenderingSupport() {
-		return renderingSupport;
+		return this;
+	}
+
+	@Override
+	public void render(
+			SqlAppender sqlAppender,
+			String functionName,
+			List<SqlAstNode> sqlAstArguments,
+			SqlAstWalker walker,
+			SessionFactoryImplementor sessionFactory) {
+		final boolean useParens = useParenthesesWhenNoArgs || !sqlAstArguments.isEmpty();
+
+		sqlAppender.appendSql( functionName );
+		if ( useParens ) {
+			sqlAppender.appendSql( "(" );
+		}
+
+		boolean firstPass = true;
+		for ( SqlAstNode sqlAstArgument : sqlAstArguments ) {
+			if ( !firstPass ) {
+				sqlAppender.appendSql( ", " );
+			}
+			sqlAstArgument.accept( walker );
+			firstPass = false;
+		}
+
+		if ( useParens ) {
+			sqlAppender.appendSql( ")" );
+		}
 	}
 
 	@Override
